app/requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
python-decouple
psycopg2-binary
gunicorn
drf-yasg
requests
twilio
djangorestframework-simplejwt[crypto]